refactor: extract environment variables
- 환경변수 분리 리팩토링
