+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/_core/config/EnvConfig.java
@@ -1,0 +1,22 @@
+package com.kakao.sunsuwedding._core.config;
+
+import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
+import org.springframework.core.env.PropertiesPropertySource;
+import org.springframework.core.env.PropertySource;
+import org.springframework.core.io.support.EncodedResource;
+import org.springframework.core.io.support.PropertySourceFactory;
+
+import java.io.IOException;
+import java.util.Objects;
+import java.util.Properties;
+
+public class EnvConfig implements PropertySourceFactory {
+    @Override
+    public PropertySource<?> createPropertySource(String name, EncodedResource resource) throws IOException {
+        YamlPropertiesFactoryBean factoryBean = new YamlPropertiesFactoryBean();
+        factoryBean.setResources(resource.getResource());
+        Properties properties = factoryBean.getObject();
+        assert properties != null;
+        return new PropertiesPropertySource(Objects.requireNonNull(resource.getResource().getFilename()), properties);
+    }
+}
