+++ sunsu-wedding/src/test/java/com/kakao/sunsuwedding/match/MatchRestControllerTest.java
@@ -1,6 +1,6 @@
 package com.kakao.sunsuwedding.match;
 
-import com.kakao.sunsuwedding._core.security.SecurityConfig;
+import com.kakao.sunsuwedding._core.config.SecurityConfig;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
