+++ sunsu-wedding/src/test/java/com/kakao/sunsuwedding/payment/PaymentRestControllerTest.java
@@ -1,7 +1,7 @@
 package com.kakao.sunsuwedding.payment;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.kakao.sunsuwedding._core.security.SecurityConfig;
+import com.kakao.sunsuwedding._core.config.SecurityConfig;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
