+++ sunsu-wedding/src/test/java/com/kakao/sunsuwedding/portfolio/PortfolioControllerTest.java
@@ -1,7 +1,7 @@
 package com.kakao.sunsuwedding.portfolio;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.kakao.sunsuwedding._core.security.SecurityConfig;
+import com.kakao.sunsuwedding._core.config.SecurityConfig;
 import org.hamcrest.core.IsNull;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
